backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
cryptography==40.0.2
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==3.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   social-auth-app-django
django-cors-headers==3.14.0
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via djoser
djoser==2.2.0
    # via -r requirements.in
idna==3.4
    # via requests
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
pycparser==2.21
    # via cffi
pyjwt==2.6.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python3-openid==3.2.0
    # via social-auth-core
pytz==2023.3
    # via
    #   django
    #   djangorestframework
requests==2.30.0
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
social-auth-app-django==5.2.0
    # via djoser
social-auth-core==4.4.2
    # via social-auth-app-django
sqlparse==0.4.4
    # via django
urllib3==2.0.2
    # via requests
